Force board size to 19 by 19

+++ src/godr/frontend/sgfpainter.py
@@ -55,8 +55,8 @@ class SgfPainter(QWidget):
 
         width = s - 60  # width of the actual board
         width_per_col = width / (sgf_board_size-1)
-        r = (width_per_col / 2.1)  # radius of stones
-        coords = [(i * width_per_col + 30) for i in range(sgf_board_size)]
+        r = int(width_per_col / 2.1)  # radius of stones
+        coords = [int(i * width_per_col + 30) for i in range(sgf_board_size)]
         cmin, cmax = min(coords), max(coords)
 
         qp.translate((w - s) // 2, (h - s) // 2)  # draw board at the center
